requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
cryptography==43.0.1
redis==5.0.8
typer==0.12.5
orjson==3.10.7
pytest==8.3.2
hiredis==3.0.0
httpx==0.27.0
